dcache-chimera: limit the number of cached hsm locations in Cleaner
Motivation:

The HsmCleaner regularly fetches all HSM locations for deletion from the trash table and caches them locally for batched dispatch. When too many entries have accumulated in the trash table, the Cleaner cell may run out of memory in the attempt to store them all.
Therefore the maximum number of cached locations should be limited.

Modification:

A limit on the number of locations that will be cached at any point in time is introduced. It is configurable and should be adjusted according to the available resources.
This chunking is achieved by fetching and caching as many locations from the database as there is space left, starting from an offset that corresponds to the already fetched entries in this database iteration. When the end of the database is reached, this offset is reset in order to retry previously unsuccesful locations.

Result:

The HsmCleaner will no longer run out of memory if the trash table is too full to be held in memory.

Target: master
Closes: #2529
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/12509/
Acked-by: Tigran Mkrtchyan